Run more tests even when CCI is missing
