+++ tests/test_with_cci.py
@@ -1,15 +1,14 @@
 from pathlib import Path
-from tempfile import NamedTemporaryFile, TemporaryDirectory
+from tempfile import TemporaryDirectory
 import unittest
-from click import ClickException
 from sqlalchemy import create_engine
 
 
-from snowfakery.cli import generate_cli, eval_arg
+from snowfakery.cli import generate_cli
 
 try:
     import cumulusci
-except:
+except ImportError:
     cumulusci = False
 
 sample_mapping_yaml = Path(__file__).parent / "mapping_vanilla_sf.yml"
@@ -39,40 +38,3 @@ class Test_CLI_CCI(unittest.TestCase):
             result = list(connection.execute("select * from Account"))
             assert result[0]["id"] == 1
             assert result[0]["BillingCountry"] == "Canada"
-
-    @unittest.skipUnless(cumulusci, "CumulusCI not installed")
-    def test_mapping_file_no_dburl(self):
-        with self.assertRaises(ClickException):
-            generate_cli.main(
-                ["--mapping_file", str(sample_mapping_yaml), str(sample_yaml)],
-                standalone_mode=False,
-            )
-
-    @unittest.skipUnless(cumulusci, "CumulusCI not installed")
-    def test_mutually_exclusive(self):
-        with self.assertRaises(ClickException) as e:
-            with TemporaryDirectory() as t:
-                generate_cli.main(
-                    [
-                        str(sample_yaml),
-                        "--dburl",
-                        f"csvfile://{t}/csvoutput",
-                        "--output-format",
-                        "JSON",
-                    ],
-                    standalone_mode=False,
-                )
-        assert "mutually exclusive" in str(e.exception)
-
-        with self.assertRaises(ClickException) as e:
-            generate_cli.main(
-                [
-                    str(sample_yaml),
-                    "--cci-mapping-file",
-                    sample_mapping_yaml,
-                    "--output-format",
-                    "JSON",
-                ],
-                standalone_mode=False,
-            )
-        assert "apping file" in str(e.exception)
